+++ service/common/src/main/java/org/apache/polaris/service/context/DefaultCallContextResolver.java
@@ -24,7 +24,7 @@ import jakarta.inject.Inject;
 import java.time.Clock;
 import java.util.Map;
 import org.apache.polaris.core.PolarisCallContext;
-import org.apache.polaris.core.PolarisConfigurationStore;
+import org.apache.polaris.core.config.PolarisConfigurationStore;
 import org.apache.polaris.core.PolarisDiagnostics;
 import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.context.RealmContext;
